MatrixPilot_AUAV3: internal logfile improvements: logfile started/stopped either on flight/landing detection otherwise can be manually controlled by DIG2 to ground. imported remzibi and minim OSD modules.
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_AUAV3@2817 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libUDB/barometer.c
@@ -257,7 +257,8 @@ void ReadBarTemp_callback(boolean I2CtrxOK)
 	if (I2CtrxOK == true)
 	{
 //		byteswaparray((unsigned char*)&barData, 2);
-		ut = ((unsigned int) (unsigned int)barData[0] << 8 | (unsigned int)barData[1]);
+//		ut = ((unsigned int) (unsigned int)barData[0] << 8 | (unsigned int)barData[1]);
+		ut = (barData[0] << 8 | barData[1]);
 #ifdef TEST_WITH_DATASHEET_VALUES
 		ut = 27898;
 #endif
